pyproject.toml
[project]
name = "ask-ai-cooper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.133.1",
    "huggingface-hub>=0.36.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "transformers>=4.57.6",
    "uvicorn>=0.41.0",
]
